requirements.txt
PyMuPDF
Pillow
ebooklib
tkinterdnd2
